rename and move `folder_black_list_test.py`; split DeltaGeneratorTestCase into its own file
